Replace Subplebbit/Author Ipns Name with Subplebbit/Author Address
"Ipns Name" is not an appropriate word because "names" can also be eth addresses, bitcoin addresses, as well as crypto domains like .eth, but the crypto domains won't resolve using the TXT record "content" like regular Ipns Names. We could have also called it just "name", but I think using the same vocabulary as cryptocurrencies makes sense, people are already used to using the term "address" to describe crypto names like john.eth, but they are not used to using a random hash as a "name", so "address" seems the least confusing.